+++ requirements.txt
@@ -1,6 +1,7 @@
 attrs==20.2.0
 click==7.1.2
 Flask==1.1.2
+Flask-SQLAlchemy==2.4.4
 iniconfig==1.1.1
 itsdangerous==1.1.0
 Jinja2==2.11.2
@@ -12,5 +13,6 @@ PyJWT==1.7.1
 pyparsing==2.4.7
 pytest==6.1.2
 six==1.15.0
+SQLAlchemy==1.3.20
 toml==0.10.1
 Werkzeug==1.0.1
