--- token_menager/blueprints/token/views.py
@@ -1,9 +1,0 @@
-from flask import Blueprint
-
-
-token_manager_blueprint = Blueprint('manager', __name__)
-
-
-@token_manager_blueprint.route('/token')
-def get_token():
-    return ''
